test: src/getEndpointDiscoveryCommandPlugin.spec.ts
